modify response

+++ APP/DAL/ComponentDAL.js
@@ -614,7 +614,7 @@ select
   , pm.created_at
   , model
   , pc.processor_chipset  as chipset
-  , pc.id as chipset_id
+  , pc.id::integer as chipset_id
   , cores
   , threads
   , model_number
